Modifications to Gridforce:
1) Changed default potential file units to kcal/mol*e (volts available
w/ keyword gridforcevolts).

2) Atom charge for coupling to grid can be specified in PDB
(gridforcefile keyword). The only practical side-effect is an increase
in memory usage, because the potential grid is now saved in addition
to the pre-computed force grid currently used to calculate forces =>
memory usage is now sizeof(float) * gridpoints * 4 instead of * 3.
Soon the pre-computed force grid will be scrapped in favor of a more
elegant interpolation scheme for the potential grid which can then be
directly differentiated to find the force.

Also changed Make.depends so that NAMD compiles! Proper entries were
not entirely in place for AtomsDisInfo nor CompressPsf.