Fixed a subtle bug in the memory-reduced version.
When compressing the psf file, we need to know the type of each tuple and integrate it to the atom signature. As a result, the Parameter object is modified through the call of assign_diheral_index and assign_improper_index. Particularly, the "multiplicity" field in the dihedral_array/improper_array is changed.

When it comes to reading the compressed psf file, we will miss such changed information for the Parameter object because the above two functions will not be called. Consequently, we will get wrong values for the DIHEDRAL/IMPROPER energy output.

The way I fix it is through adding these modified "multiplicity" values in the compressed psf file. When reading the compressed psf file, we will update the Parameter object with these values.