fixed the problems in supporting extra bonds feature for memory optimized version.
The problems appeared in the previous CVS versions are caused by not differentiating "numRealBonds" and "numBonds" which are the two variables inside Molecule class. For example, the exclusions should be determined only by real bonds excluding those extra bonds. In addition, the Molecule::build_atom_status uses numRealBonds to test the status of each atom. In the mmeory optimized version, bonds information is obtained through atom signatures (the bond signature, more specifically). Obviously, the real bonds and the extra bonds cannont not be differentiated through the signatures. Therefore, one variable "isReal" is added to TupleSignature structure to indicate whether this tuple is real or is an extra tuple read from the file other than psf file. This addition will help the future implementation of other extra tuples such as angles, dihedrals and impropers.