1. Add comppile-time flags to allow users to choose whether to remove the extra copy overhead of proxy msgs (ProxyDataMsg and ProxyResultMsg).
2. Add the optimization to remove the extra copy involved in the proxy
result msg.

2. Add padding bytes to the varsize proxy msgs, so that the beginning of
CompAtom list in the ProxyDataMsg is 32-byte aligned if the underlying
Charm layer has ensured the beginning of a msg is 32-byte aligned;
the beginning of Force list in the ProxyResultVarsizeMsg is 8-byte
aligned. This is done to accord to the alignment requirement before this
optimization since the alignment will give better performance,
especially on the BG/L machine.