Attempted to optimize inner loops of grid cutoff by moving as much indexing arithmetic as possible outside of loops.  Also have two branches that tests for inner loop of length 8 (occurs about 50%). The case with the length 8 inner loop is hard coded.
Also added profiling reduction to gather statistics on the length
of the innermost loop.  Cannot yet enable both the timing and profiling
reductions together.  Fixed the reductions across ComputeMsmMgr so that
objects check in with ComputeMsmMgr and then call done when finished
with reduction.

Both reductions are disabled.  Enable by changing #define macros.