Fix modest bug in alch TI
Thermodynamic integration (TI) output has two sets of output fields inside
alchOutfile - a block average of the derivatives and a running average of the
derivatives. For some reason, the block average (which is probably never
actually used for analysis) was NOT being reset after alchEquilSteps was
completed. This caused disagreement between the two averages after the first
sampling interval, even though they should be exactly the same.

Change-Id: Ic7c8f00389aaa047e2b5df7b49c43dd66bf598cb